cleanup and refactoring to use span for Steam and Microsoft

+++ libNOM.io/Enums/DifficultyPresetTypeEnum.cs
@@ -1,10 +1,10 @@
 ﻿namespace libNOM.io.Enums;
 
 /// <summary>
-/// Specifies available game modes.
+/// Specifies all available difficulties (PresetGameModeEnum was used for this before 4.00).
 /// </summary>
 /// <seealso cref="libMBIN\Source\NMS\GameComponents\GcDifficultyPresetType.cs"/>
-public enum DifficultyPresetTypeEnum : byte
+public enum DifficultyPresetTypeEnum : uint
 {
     Invalid,
     Custom,
